# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virtigia Map Editor API",
        "description": "Token do autoryzacji wygenerujesz w panelu użytkownika: Profil -> Tokeny API.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/base-npcs": {
            "get": {
                "tags": [
                    "Base NPC"
                ],
                "summary": "Pobranie listy bazowych NPC",
                "operationId": "getBaseNpcsList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginowana lista bazowych NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiBaseNpcListCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
        "/api/v1/base-npcs/{baseNpcId}": {
            "get": {
                "tags": [
                    "Base NPC"
                ],
                "summary": "Pobranie pojedynczego bazowego NPC",
                "operationId": "getBaseNpcById",
                "parameters": [
                    {
                        "name": "baseNpcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły bazowego NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiBaseNpcDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "Bazowy NPC nie został znaleziony"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
        "/api/v1/maps": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "summary": "Pobranie listy map (bez ciężkiego pola col)",
                "operationId": "getMapsList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Liczba rekordów na stronę (1-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginowana lista map",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\MapListCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
        "/api/v1/maps/{mapId}": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "summary": "Pobranie pojedynczej mapy",
                "operationId": "getMapById",
                "parameters": [
                    {
                        "name": "mapId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły mapy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\MapDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "Mapa nie została znaleziona"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Maps"
                ],
                "summary": "Edycja mapy (name, col, pvp)",
                "operationId": "updateMap",
                "parameters": [
                    {
                        "name": "mapId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "pvp": {
                                        "type": "integer"
                                    },
                                    "col": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mapa została zaktualizowana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\MapDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "Mapa nie została znaleziona"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "API Token",
                "description": "Token do autoryzacji wygenerujesz w panelu użytkownika: Profil -> Tokeny API."
            },
            "worldHeader": {
                "type": "apiKey",
                "description": "Wymagany nagłówek świata. Dostępne wartości: retro, classic, legacy, demo.",
                "name": "X-World",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": [],
            "worldHeader": []
        }
    ]
}